images in req source as component
