+++ compliance-web/src/components/App/Inspections/Profile/Requirements/RequirementSource/RequirementSourceModal.tsx
@@ -1,15 +1,10 @@
 import {
   Box,
-  Button,
   DialogContent,
-  IconButton,
   Stack,
-  Typography,
-  Popover,
-  CircularProgress,
 } from "@mui/material";
 import type React from "react";
-import { useCallback, useEffect, useMemo, useRef, useState } from "react";
+import { useEffect, useMemo, useRef, useState } from "react";
 import * as yup from "yup";
 import { FormProvider, useForm, useWatch } from "react-hook-form";
 import { yupResolver } from "@hookform/resolvers/yup";
@@ -29,14 +24,8 @@ import { useInspectionOrdersProjectwiseData } from "@/hooks/useInspectionOrders"
 import { CaseFile } from "@/models/CaseFile";
 import { useCaseFileByNumber } from "@/hooks/useCaseFiles";
 import { formatAuthorization } from "@/utils/appUtils";
-import {
-  CloseRounded,
-  ImageOutlined,
-  UploadRounded,
-} from "@mui/icons-material";
-import { BCDesignTokens } from "epic.theme";
-import { useImageUpload } from "@/hooks/useImageUpload";
-import { RequirementImage, UploadedImageFile } from "@/models/Image";
+import { RequirementImage } from "@/models/Image";
+import ImagesRequirementSource from "../Images/ImagesRequirementSource";
 
 type RequirementSourceModalProps = {
   onSubmit: (data: RequirementSourceFormData) => void;
@@ -261,71 +250,6 @@ const RequirementSourceModal: React.FC<RequirementSourceModalProps> = ({
     hasUserEditedTitle.current = true;
   };
 
-  const [previewAnchorEl, setPreviewAnchorEl] = useState<HTMLElement | null>(
-    null
-  );
-  const [previewUrl, setPreviewUrl] = useState<string | null>(null);
-  const previousPreviewUrlRef = useRef<string | null>(null);
-
-  const handlePreviewEnter = (
-    event: React.MouseEvent<HTMLElement>,
-    file: RequirementImage
-  ) => {
-    setPreviewAnchorEl(event.currentTarget);
-    if (previousPreviewUrlRef.current) {
-      URL.revokeObjectURL(previousPreviewUrlRef.current);
-      previousPreviewUrlRef.current = null;
-    }
-    const url = file.url ?? "";
-    previousPreviewUrlRef.current = url;
-    setPreviewUrl(url);
-  };
-
-  const handlePreviewLeave = () => {
-    setPreviewAnchorEl(null);
-    if (previousPreviewUrlRef.current) {
-      URL.revokeObjectURL(previousPreviewUrlRef.current);
-      previousPreviewUrlRef.current = null;
-    }
-    setPreviewUrl(null);
-  };
-
-  useEffect(() => {
-    return () => {
-      if (previousPreviewUrlRef.current) {
-        URL.revokeObjectURL(previousPreviewUrlRef.current);
-        previousPreviewUrlRef.current = null;
-      }
-    };
-  }, []);
-
-  const onSuccess = useCallback((uploadedFile: UploadedImageFile) => {
-    const requirementImage: RequirementImage = {
-      relative_url: uploadedFile.relative_url,
-      url: uploadedFile.presigned_url,
-      original_file_name: uploadedFile.fileName,
-    };
-    setUploadedImages((prevImages) => [...prevImages, requirementImage]);
-  }, []);
-
-  const { mutate: uploadImage, isPending } = useImageUpload(onSuccess);
-
-  const handleUploadImage = () => {
-    const input = document.createElement("input");
-    input.type = "file";
-    input.accept = "image/*";
-    input.onchange = (event) => {
-      const file = (event.target as HTMLInputElement).files?.[0];
-      if (file) {
-        uploadImage({
-          inspectionId: caseFile.id,
-          fileName: file.name,
-          file: file,
-        });
-      }
-    };
-    input.click();
-  };
 
   return (
     <>
@@ -498,93 +422,11 @@ const RequirementSourceModal: React.FC<RequirementSourceModalProps> = ({
                     />
                   </Box>
                 </Box>
-                <Stack
-                  direction={"row"}
-                  gap={2}
-                  alignItems={"center"}
-                  sx={{ my: 2 }}
-                >
-                  <Button
-                    color="secondary"
-                    size="small"
-                    onClick={handleUploadImage}
-                    startIcon={<UploadRounded />}
-                  >
-                    Upload Image
-                  </Button>
-                  <Typography variant="caption">
-                    Uploaded files display as links here, but will be shown as
-                    images in the report.
-                  </Typography>
-                </Stack>
-                {isPending && <CircularProgress size={20} />}
-                {!isPending && (
-                  <Box
-                    sx={{
-                      maxHeight: "136px",
-                      overflow: "scroll",
-                    }}
-                  >
-                    {uploadedImages.map((item, index) => (
-                      <Box
-                        key={index}
-                        sx={{
-                          mb: 2,
-                          display: "flex",
-                          alignItems: "center",
-                          justifyContent: "space-between",
-                          gap: 1,
-                          border: "1px solid",
-                          borderColor: BCDesignTokens.surfaceColorBorderDefault,
-                          borderRadius: BCDesignTokens.layoutBorderRadiusMedium,
-                          px: "0.5rem",
-                          cursor: "pointer",
-                        }}
-                        onMouseEnter={(e) => handlePreviewEnter(e, item)}
-                        onMouseLeave={handlePreviewLeave}
-                      >
-                        <Box
-                          display="flex"
-                          alignItems="center"
-                          gap={1}
-                          color={BCDesignTokens.typographyColorLink}
-                        >
-                          <ImageOutlined />
-                          {item.original_file_name}
-                        </Box>
-                        <IconButton>
-                          <CloseRounded />
-                        </IconButton>
-                      </Box>
-                    ))}
-                    <Popover
-                      open={Boolean(previewAnchorEl)}
-                      anchorEl={previewAnchorEl}
-                      onClose={handlePreviewLeave}
-                      anchorOrigin={{ vertical: "top", horizontal: "center" }}
-                      transformOrigin={{
-                        vertical: "bottom",
-                        horizontal: "center",
-                      }}
-                      sx={{ pointerEvents: "none" }}
-                    >
-                      {previewUrl && (
-                        <Box
-                          component="img"
-                          src={previewUrl}
-                          alt="Preview"
-                          sx={{
-                            maxWidth: 420,
-                            maxHeight: 420,
-                            objectFit: "contain",
-                            display: "block",
-                            padding: "4px",
-                          }}
-                        />
-                      )}
-                    </Popover>
-                  </Box>
-                )}
+                <ImagesRequirementSource
+                  uploadedImages={uploadedImages}
+                  setUploadedImages={setUploadedImages}
+                  inspectionId={caseFile.id}
+                />
               </Box>
             </Stack>
           </DialogContent>
